+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 2.2.1
 
 - Remove tap-highlight for checkbox, radio and switch on mobile devices
+- Select: Fixed disabled icon color